Fixed a calendar bug
If you click on a date from previous month, it changes month to display previous one. But the bug was - if the current date is higher than some date from previous month, it would apply the functionality like it is a past date from current month. That would make it unclickable. 

Example: 
It is 29th April. If you're looking at April calendar, previous month has dates from 27 through 31 displayed, from Sunday to Thursday.
Instead of all March days to be clickable and change calendar to March, the ones lower than 29 couldn't be clicked. It showed them in red (current color for days in the past that couldn't be clicked), as opposed to 29, 30, 31 March that were gray (current color for days from previous month that can be clicked to change calendar to that month). 

Added the conditional that index of those days must not be higher than first day of the month ( getDay()) in order to be considered a day in the past (red date).